[BUG FIXES] - Text in Bottom Sheets - Crash with login on different device - ConstraintLayout to MaterialCardView

+++ app/src/main/java/com/nowiwr01/stop_smoking/presentation/main/smoke_info/SmokeInfoViewsController.kt
@@ -49,7 +49,7 @@ class SmokeInfoViewsController(
         currency = viewModel.currency.value!!.symbol,
         cigarettesPerDay = binding.cigarettesPerDay.text.toString().toInt(),
         cigarettesPerPack = binding.cigarettesPerPack.text.toString().toInt(),
-        cigarettesPackCost = binding.onePackCost.text.toString().toInt()
+        cigarettesPackCost = binding.onePackCost.text.toString().toDouble()
     )
 
     fun setTextChangedCallback(inputFields: List<EditText>) {
